requirements.txt
Django==5.2.7
djangorestframework==3.16.1
pymongo==4.15.3
